Renamed Sconfig to TerasliceCliConfig
