+++ packages/teraslice-cli/cmds/aliases/update.js
@@ -2,7 +2,7 @@
 'use console';
 
 const reply = require('../lib/reply')();
-const Sconfig = require('../lib/sconfig');
+const TerasliceCliConfig = require('../lib/teraslice-cli-config');
 const appCli = require('../lib/app-cli');
 const clusterUrlCli = require('../lib/cli/cluster-url');
 const cmdCli = require('./lib/cmd-cli');
@@ -18,8 +18,13 @@ exports.builder = (yargs) => {
 };
 
 exports.handler = (argv, _testFunctions) => {
-    const cliConfig = new Sconfig(argv);
+    const cliConfig = new TerasliceCliConfig(argv);
     const libAliases = _testFunctions || require('./lib')(cliConfig);
+
+    if (!(cliConfig.args.cluster_alias && cliConfig.args.cluster_url)) {
+        reply.fatal('You must specify both a cluster alias and cluster URL');
+    }
+
     return libAliases.update()
         .catch(err => reply.fatal(err.message));
 };
